refactor: rename package timer to repeat
I believe repeat is a better name for this functionality.
