+++ internal/server/server.go
@@ -32,9 +32,9 @@ import (
 	"github.com/infrahq/infra/internal"
 	"github.com/infrahq/infra/internal/certs"
 	"github.com/infrahq/infra/internal/logging"
+	"github.com/infrahq/infra/internal/repeat"
 	"github.com/infrahq/infra/internal/server/data"
 	"github.com/infrahq/infra/internal/server/models"
-	timer "github.com/infrahq/infra/internal/timer"
 	"github.com/infrahq/infra/pki"
 	"github.com/infrahq/infra/secrets"
 )
@@ -170,7 +170,7 @@ func configureTelemetry(db *gorm.DB) error {
 		return err
 	}
 
-	timer.Start(context.TODO(), 1*time.Hour, func(context.Context) {
+	repeat.Start(context.TODO(), 1*time.Hour, func(context.Context) {
 		if err := tel.EnqueueHeartbeat(); err != nil {
 			logging.S.Debug(err)
 		}
